feat: Complete Phase 6 meteor shower refactor with unified render pipeline
- Implement single updateAll() entry point replacing 6+ individual WASM calls
- Add comprehensive JavaScript fallback for full WASM/JS compatibility
- Create differential rendering system with dirty flags and temporal coherence
- Fix negative particle/meteor radius errors with proper bounds checking
- Add canvas resize support for both WASM and JavaScript pipelines
- Implement zero-copy memory management with TypedArray views
- Move all legacy code to ms-legacy/ for clean architecture
- Add proper data extraction from packed render buffers
- Reduce WASM boundary crossings from 6+ to 1 per frame for better performance

This completes the Phase 6 unified render pipeline as outlined in the
meteor shower refactor guide, providing a robust foundation for high-
performance meteor effects with automatic fallback capabilities.
