+++ ms-legacy/wasm/src/particle_pool.rs
